issue-1541: moved TFileRingBuffer implementation to .cpp
